5.2 form db update

+++ db/migrate/20170812115947_add_price_to_products.rb
@@ -1,5 +1,5 @@
 class AddPriceToProducts < ActiveRecord::Migration[5.1]
   def change
-    add_column :products, :price, :float
+    add_column :products, :price, :integer
   end
 end
